added confirmation prompt swal

+++ src/App.js
@@ -283,7 +283,19 @@ function PublicChat({ user }) {
                           <AiIcons.AiOutlineDelete
                           className="delete-button"
                           onClick={() => 
-                            deleteMessage(msg.id)
+                            Swal.fire({
+                              title: 'Are you sure?',
+                              text: "You won't be able to revert this!",
+                              icon: 'warning',
+                              showCancelButton: true,
+                              confirmButtonColor: '#3085d6',
+                              cancelButtonColor: '#d33',
+                              confirmButtonText: 'Yes, delete it!'
+                            }).then((result) => {
+                              if (result.isConfirmed) {
+                                deleteMessage(msg.id)
+                              }
+                            })
                           } />
                         )}
                         {messageTime}
